[dotnet] Enable generation of runtime configuration file Set the GenerateRuntimeConfigurationFiles property to be true to avoid warnings at build time for users + add test for change.
Diving deeper into the fix...
- GRCF == GenerateRuntimeConfigurationFiles property
- This warning only occurs with .NET apps which is why the property is only
updated in the dotnet directory and not msbuild (legacy code)
- Diving deeper into the fix, after examining the binlog (in #17543),
it was found that the GRCF was contingent upon the
HasRuntimeOutput property, which is only defined for executable projects.
And in this case, the user's project output type is library thus both the
RuntimeOutput and consequently GRCF properties were not enabled.
- By setting the GRCF to true we can address the original warning of concern
while ensuring the rest of the projects's behavior is not altered in mysterious
ways (i.e. by touching the RuntimeOutput property or the project output
type instead, these changes could have extraneous effects).

Fixes #17543