[src] Fix containing namespace/framework for AVCustomRoutingControllerDelegate.
The AVCustomRoutingControllerDelegate protocol was incorrectly placed in the
AVKit framework, instead of the AVRouting framework where it currently belongs
(this might have been a change between betas at some point).

So move AVCustomRoutingControllerDelegate to the AVRouting namespace, but only
in XAMCORE_5_0, because otherwise it would be a breaking change.

In the meantime, special case the AVCustomRoutingControllerDelegate type so
that we correctly link with the AVRouting framework when it's used.