[xharness] Change project generation to not add absolute paths.
Change project generation to not write out absolute paths, because we're going to
generate projects on one machine, and then build and execute them on another, and
the absolute path might be wrong.

Instead use the 'RootTestsDirectory' property in the generated project files, and
then set the property to the correct absolute path when the tests are executed on
the target machine.

This required a few changes all over the place, because in some cases we read/evaluate
project files in order to figure out what to do with them, and some of that code
needs updating to cope with the fact that it might run into paths with '$(RootTestsDirectory)'
in them, and these must be resolved to a absolute path before trying to find files
on disk.