[ci] Add parameter to disable classic signing
A new parameter has been added to allow the unified pipeline to skip
classic signing jobs.

The jobs that would sign and notarize the custom workload .pkg and .zip
files have been removed, as we do not ship these anywhere and do not
need to wait on signing for them.  The `dotnet-signed` artifact has been
removed as a result, and the `not-signed-package` artifact is used in
its place.