[monotouch-test] Adjust ImageCaptioningTest.GetCaption to don't care if we get an error or not. (#18267)
This check is passing a remote url to an API that wants a url to a local file
(file:// url), and this is supposed to fail (it's not a network hiccup if it
fails).

However, sometimes it doesn't fail... for unknown reasons. So instead change
the test's expectations so that it passes whether using a remote url fails
with an error or not.