[DevOps] Provide a script that will allow use to debug the expanded yaml.
This pr provides a script that will allow to create a dry run in vsts
and will return the expanded yaml after the yaml compiler has been
executed, this will be written in the given output file.

Usage:

Basic usage: Query the expanded template from the ci pipeline in main:

```pwsh
./preview_yaml.ps1 -AccessToken $pat -OutputFile ./full.yaml
```

Query the expansion in the pr pipeline:
```pwsh
./preview_yaml.ps1 -AccessToken $pat -OutputFile ./full.yaml -Pipeline "pr"
```

Query the expansion in the ci pipeline for a diff branch than main:
```pwsh
./preview_yaml.ps1 -AccessToken $pat -OutputFile ./full.yaml -Branch "trigger-issues"
```

The output file is ALLWAYS overriden.