[dotnet-linker] Fix detecting when we must call the object-based Runtime.GetNSObject
We can't call the generic Runtime.GetNSObject<T> method when we don't know the final
type of the returned object, which happens for open types. However, we must always
call the generic Runtime.GetNSObject<T> when we know the final type, so that we have
the correct object type on the evaluation stack.

Fix this by properly detecting open types and only calling Runtime.GetNSObject in
that case.