[src] Add generated file listing frameworks to git to simplify build on Windows.
This makes it easier to run some tests on Windows, because the tests reference
this file and by having it checked in it doesn't need to be generated (which
we currently do using make, which isn't trivial on Windows).

Also adjust the generated output slightly to make it easier on the eyes.