[dotnet] Add targets to compute mlaunch arguments for installing and launching mobile apps. Fixes #18359.
Add public targets to compute the mlaunch command lines for installing and launching mobile apps.

These new targets are:

* ComputeMlaunchInstallArguments
* ComputeMlaunchRunArguments

As part of this change, also create a few new public properties:

* MlaunchPath
* MlaunchRunArguments
* MlaunchInstallArguments
* MlaunchRunScript
* MlaunchInstallScript

If the *Script variables are set, the corresponding target will create a
script file with the path to mlaunch + the corresponding arguments.

Otherwise, it's also possible to get the arguments directly from the build
log.

Fixes https://github.com/xamarin/xamarin-macios/issues/18359.