[dotnet] Add a property to opt-out of the _CopyLocalBindingResources logic easily.
It seems this target has more problems than at first I thought, so make it
easier to opt-out of it by just setting a property in the csproj.

More investigation is needed, but I'm keeping the target on by default for
now, since it solves a real-world problem as well.