[msbuild] Fix check for Link metadata by checking for both null and empty string. Fixes #12783.
A task item's GetMetadata method will return an empty string if the metadata
isn't available, so make sure to check for both null (which shouldn't really
happen, but just to be safe) and an empty string when detecting whether the
Link metadata is set or not.

Fixes https://github.com/xamarin/xamarin-macios/issues/12783.