[UIKit] Fix UIKit selector on TvOS
This API was missed because the old sim intro tests are not running and
we missed it. This was found while working on xcode15. The tvOS attr is
missing which defaults to the min version of tvOS when it should be 15.

Added a compat implementation so that the API remains.