[dotnet] Link Mono and Xamarin statically in Mac Catalyst by default. Fixes #14686.
It's possible to create a provisioning profile for Mac Catalyst that doesn't
allow dylibs in the app. It seems a significant number of people run into this
problem when publishing their apps, so avoid it by linking Mono and Xamarin
statically by default instead.

The downside is that build time might increase a little bit.

An upside however is that the app size might decrese somewhat.

Fixes https://github.com/xamarin/xamarin-macios/issues/14686.