[Generator] Ensure that selectors fields do not have overlapping names. Fixes #18645
The way the generator names private fields for selectors results on not
compilable code when there are two selectoors A and B such that selector
B == AHandle.

That is, if selector A is "user", selector B will be
"userHandle". This happens because Handle is quite a common postfix in
security when working with user accounts (userHandler and others) and
we need to ensure that the Handle postfix added by the generator is unique.

As stated in the bug we could try to fix this by keeping track of the variables in the context
in wich the generator is creating code. The problem with such approach
is that it will make very hard to predict the name of the fields making
the manual code harder to write. The PR contains examples of such manual
code.

To fix the situation we have moved from using Handle to XHandle as a
postfix in the field names whihc reduces the chances of finding a
similar corner case in the future.

A test has been added to show the case in which we found the bug.

Fixes #18645