[devops] Remove option to disable .NET from CI.
It's rather rare now that we need to disable .NET from the CI, so just remove
the option. It can still be disabled directly in the code when needed.

Also this fixes an issue where we'd build with .NET enabled even if disabled
in the code, because we'd always pass --enable-dotnet to configure.