[net8.0] [msbuild] Re-aot referencing assemblies. Fixes #17708. (#18518)
If an assembly changes, then we must AOT compile that assembly again (which we already
did), in addition to any assembly that references the modified assembly (which we
didn't do).

So rework the AOTCompile target: remove the Inputs and Outputs (because the dependency
tracking is too complicated for MSBuild to resolve), and instead move the logic to
detect if an assembly must be AOT-compiled again into the AOTCompile task.

This is a port of the [original fix for main][1] to .NET 8, where we have to
take the dedup assembly into account too: if any assembly has changed, then we
must re-aot the dedup assembly as well.

Fixes https://github.com/xamarin/xamarin-macios/issues/17708.

[1]: https://github.com/xamarin/xamarin-macios/pull/18509

---------

Co-authored-by: Alex Soto <alex@alexsoto.me>