[release-test/multi-targetting-7.0.1xx-xcode14] [dotnet] Parameterize the pack names.
We're going to change the pack names to support multi-targetting, so ahead of
the pack name change I'm changing the existing logic to use a variable for the
pack name in most places (this will make the rename much easier and simpler).

These changes should have no effect by themselves.