[dotnet] Copy assemblies to AOT to a new directory if they're in a directory with a colon. Fixes #14904.
The AOT compiler fails we pass a MONO_PATH with a colon (because Mono supports
multiple paths in MONO_PATH, separated by colon...). So if we get assemblies
in such a directory (which happens for remote builds, when
IntermediateOutputPath is an absolute (Windows) path), copy them all to a
different (temporary) directory and use that instead.

It's slower, but at least it works.

Fixes https://github.com/xamarin/xamarin-macios/issues/14904.