[tools] Store list of trampolines even if it's empty. Fixes #18570.
This is a bit non-obvious, but we later use the presence of such a stored list
to determine if we need to generate the __Registrar__ type in the assembly.

The problem is that we need to generate the __Registrar__ type even if no
trampolines were created, because we use it for numerous other purposes as
well (type lookup for instance).

Fixes https://github.com/xamarin/xamarin-macios/issues/18570.