[devops] Tell 'dotnet test' to use the correct 'dotnet' executable when running the generator tests.
'dotnet test' has this very nonintuitive behavior where it will execute the
tests with the 'dotnet' executable found in PATH, not the 'dotnet' executable
we're executing.

This matters when we're using a preview version of .NET: the system does not
have that preview version installed, so executing the tests with our locally
built .NET doesn't work, because it just switches to using the system .NET,
which doesn't have the version we need.

So use a settings file to tell .NET to not go looking for .NET in any other
.NET location than the .NET location we want it to use.