[bgen] Add ToArray and ToFlags extension methods for strongly typed [Flags] enums. (#18925)
We have a few different implementations of how to compute a set or array of NSString
constants from an enum value, and also computing an enum value from a set or array
of NSString constants.

So add support to the generator for generating a ToArray and a ToFlags method for
these conversions, and remove all the manual code.

This has a few advantages:

* The generated code is automatically updated with new enum fields (not all previous
  implementations were).
* Reduces code duplication.
* Reduces manual code.

Additionally, when computing between a group of constants and the corresponding flags,
we have to handle any missing constants or flags gracefully, otherwise the code won’t
be forward compatible.

Example (from bug #18833):

* App developer writes code to fetch the AVCaptureMetadataOutput.AvailableMetadataObjectTypes
  property, which is an enum value of flags.
* App developer ships their app in the app store.
* Apple releases a new iOS update, adding new values and constants to the enum.
* The native [AVCaptureMetadataOutput availableMetadataObjectTypes] method returns
  these new constants.
* This means that if the AVCaptureMetadataOutput.AvailableMetadataObjectTypes property
  throws an exception if encountering unknown constants, our API wouldn’t be forward
  compatible.
* The correct solution here is that AVCaptureMetadataOutput.AvailableMetadataObjectTypes
  must ignore any constants it doesn’t know about.

Fixes https://github.com/xamarin/xamarin-macios/issues/18833.