[ObjCRuntime] Remove unnecessary macOS code from .NET. (#18967)
In Xamarin.Mac we have some code to support unusual ways of creating app
bundles while including Xamarin.Mac, and also dynamic loading of the Mono
runtime.

In .NET, we don't support these unusual app bundle creating logic for any
platform, and the Mono runtime isn't even an option for macOS, so exclude some
of the corresponding code from .NET.

This fixes a few warnings in NativeAOT about this code not being trimmer-safe.